move Mode to constants
