+++ frontend/src/appConstants.ts
@@ -1,4 +1,9 @@
-const rows = 10;
-const columns = 10;
-
-export { rows, columns };
+export const rows = 10;
+export const columns = 10;
+export enum Mode {
+  Draw = 'draw',
+  Erase = 'erase',
+  EnterQuestions = 'enterQuestions',
+  Answer = 'answer',
+  Puzzle = 'puzzle',
+}
